moves cathryn_images into images folder
